+++ src/main/kotlin/com/maedjyukghoti/tictactoe/logic/players/Player.kt
@@ -1,6 +1,7 @@
 package com.maedjyukghoti.tictactoe.logic.players
 
 import com.github.michaelbull.result.*
+import com.maedjyukghoti.tictactoe.AppState.Game
 import com.maedjyukghoti.tictactoe.UserIntent
 import com.maedjyukghoti.tictactoe.logic.*
 
@@ -11,7 +12,7 @@ sealed interface Player {
         private val playerInfo: PlayerInfo,
         private val botStrategy: BotStrategy
     ) : Player {
-        fun getAction(gameState: GameState): UserIntent {
+        fun getAction(gameState: Game): UserIntent {
             return botStrategy.getCoordinates(gameState, playerInfo)
                 .fold(
                     success = { coordinates -> UserIntent.Move(coordinates) },
